[#4528] Make ChannelHandlerAppender sub-types compatible with previous ChannelDuplexHandler sub-types
Motivation:

ChannelHandlerAppender did remove itself from the pipeline and so changed the sematic of HttpClientCodec and HttpServerCodec compared to 4.0 in which a user could just remove one of these and so remove all handlers that where part of the codec. This leads to an API-breakage when upgrade from 4.0 to 4.1

Modifications:

- Don't remove ChannelHandlerAppender by default from pipeline
- When a ChannelHandlerAppender is removed from the pipeline by a user it will also remove all the handler that where append as part of the ChannelAppender

Result:

Remove HttpClientCodec or HttpServerCodec from the pipeline has the same sematic as in 4.0.